Delete fishoutofwater-sentient-droid.md
This character concept is problematic, after some thought. The question of droids as property lends itself too easily to situations that echo slavery too closely. I'm nixxing this playbook as a result, and will have to re-evaluate the whole  "Fish out of water" concept to ensure the other characters aren't also problematic in some way.